better support for kmz in upload
Previously, kmz was unpacked early in the process and alongside
any other uploaded files. This resulted in the layer/featuretype
name being 'doc' and makes tracking any related files more
difficult.

- add unpack method to DataFormat to support custom unpacking
- implement unpack for KMLFileFormat
- format test for unpack
- rest endpoint test for kmz and correct name

+++ src/extension/importer/core/src/main/java/org/geoserver/importer/Directory.java
@@ -88,14 +88,22 @@ public class Directory extends FileData {
     }
 
     public void unpack(File file) throws IOException {
-        //if the file is an archive, unpack it
-        VFSWorker vfs = new VFSWorker();
-        if (vfs.canHandle(file)) {
-            LOGGER.fine("unpacking " + file.getAbsolutePath() + " to " + this.file.getAbsolutePath());
-            vfs.extractTo(file, this.file);
+        DataFormat fileFormat = DataFormat.lookup(file);
+        // if unrecognized, try to unpack, otherwise skip and let the
+        // format itself do what it needs. This is important for KMZ as
+        // we want to use the kmz basename for the layer.
+        if (fileFormat == null) {
+            //if the file is an archive, unpack it
+            VFSWorker vfs = new VFSWorker();
+            if (vfs.canHandle(file)) {
+                LOGGER.fine("unpacking " + file.getAbsolutePath() + " to " + this.file.getAbsolutePath());
+                vfs.extractTo(file, this.file);
 
-            LOGGER.fine("deleting " + file.getAbsolutePath());
-            file.delete();
+                LOGGER.fine("deleting " + file.getAbsolutePath());
+                file.delete();
+            }
+        } else {
+            fileFormat.unpack(file);
         }
     }
     
@@ -383,18 +391,15 @@ public class Directory extends FileData {
         }
     }
 
-    public void accept(FileItem item) throws Exception {
+    public void accept(FileItem item) throws IOException {
         File dest = child(item.getName());
-        item.write(dest);
-
         try {
-            unpack(dest);
-        } 
-        catch (IOException e) {
-            // problably should delete on error
-            LOGGER.warning("Possible invalid file uploaded to " + dest.getAbsolutePath());
-            throw e;
+            item.write(dest);
+        } catch (Exception ex) {
+            throw new IOException("Error writing uploaded file");
         }
+
+        unpack(dest);
     }
     
     public void archive(File output) throws IOException {
